generate specs at build time
