+++ Dockerfile
@@ -41,6 +41,11 @@ ENV PATH="/root/.local/bin:${PATH}"
 RUN rm -rf package.json yarn.lock .babelrc webpack.config.js requirements.txt
 COPY --from=frontend-build /srv/static static
 
+# Build specs json file 
+ARG PRIVATE_KEY
+ARG PRIVATE_KEY_ID
+RUN python3 -m webapp.build_specs
+
 # Set git commit ID
 ARG BUILD_ID
 ENV TALISKER_REVISION_ID "${BUILD_ID}"
