+++ package.json
@@ -10,7 +10,8 @@
     "test": "yarn run lint-python && yarn run test-python",
     "test-python": "python3 -m unittest discover tests",
     "test-js": "jest --silent",
-    "test-coverage-js": "jest --coverage --silent"
+    "test-coverage-js": "jest --coverage --silent",
+    "build-specs": "python3 -m webapp.build_specs"
   },
   "dependencies": {
     "@canonical/react-components": "0.38.0",
